requirements.txt
Flask==3.0.2
Pillow==10.2.0
PyPDF2==3.0.1
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0 